+++ src/commands/validate.io.ts
@@ -1,10 +1,11 @@
 import {
-  getRuleForResult,
+  getRuleForResultV2,
   Resource,
   ValidationResponse,
 } from "@monokle/validation";
 import groupBy from "lodash/groupBy.js";
 import { B, C, S, Screen } from "../utils/screens.js";
+import { ConfigData } from "../utils/config.js";
 
 export const success = () => `
 ${S.success} All resources are valid.
@@ -42,20 +43,20 @@ export const displayInventory = (allResources: Resource[]) => {
 export const failure = (response: ValidationResponse) => {
   const screen = new Screen();
 
-  const allResults = response.runs.flatMap((r) => r.results);
-  const groupedResults = groupBy(allResults, (r) => {
-    const location = r.locations[1]?.logicalLocations?.[0]?.fullyQualifiedName;
+  const allResultsData = response.runs.flatMap((r, i) => r.results.map(result => ({result, run: i})));
+  const groupedResults = groupBy(allResultsData, (r) => {
+    const location = r.result.locations[1]?.logicalLocations?.[0]?.fullyQualifiedName;
     return location ?? "unknown";
   });
 
-  for (const [location, results] of Object.entries(groupedResults)) {
+  for (const [location, resultsData] of Object.entries(groupedResults)) {
     screen.line(C.bold(location));
 
-    for (const result of results) {
-      const color = result.level === "error" ? C.red : C.yellow;
-      const icon = result.level === "error" ? S.error : S.warning;
-      const rule = getRuleForResult(response, result);
-      const message = result.message.text;
+    for (const item of resultsData) {
+      const color = item.result.level === "error" ? C.red : C.yellow;
+      const icon = item.result.level === "error" ? S.error : S.warning;
+      const rule = getRuleForResultV2(response.runs[item.run], item.result);
+      const message = item.result.message.text;
       screen.line(`${color(`[${icon} ${rule.name}]`)} ${message}`);
     }
 
@@ -78,6 +79,31 @@ export const failure = (response: ValidationResponse) => {
   screen.line(
     B(
       ` ${icon} ${validationCount} misconfigurations found. (${errorCount} errors)`,
+      {
+        padding: 1,
+        dimBorder: true,
+      }
+    )
+  );
+
+  return screen.toString();
+};
+
+export const configInfo = (configData: ConfigData) => {
+  let configInfo = '';
+  if (configData.isFrameworkBased) {
+    configInfo = `${C.bold(configData.framework)} framework based policy`;
+  } else if (configData.isRemote) {
+    configInfo = `remote policy from ${C.bold(configData.remoteParentProject?.name ?? 'unknown')} project. It can be adjusted on ${configData.remoteParentProject?.remoteUrl}`;
+  } else {
+    configInfo = `local policy from ${C.bold(configData.path)} file`;
+  }
+
+  const screen = new Screen();
+
+  screen.line(
+    B(
+      ` ${S.info} Validating using ${configInfo}.`,
       {
         padding: 1,
         dimBorder: true,