+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # @monokle/cli
 
+## 0.5.0
+
+### Major Changes
+
+- Introduced integration with [Monokle Cloud](https://app.monokle.com/):
+  - Introduced `monokle login` command.
+  - Introduced `monokle logout` command.
+  - Introduced `monokle whoami` command.
+  - Introduced `--api-token` (`-t`) argument for `monokle validate` and `monokle login` commands to authenticate easily in CI/CD scenarios.
+- Introduced `monokle init` command to allow easy bootstrapping of local config [#4](https://github.com/kubeshop/monokle-cli/issues/4).
+- Introduced `monokle config show` command to preview configuration used for validation.
+
+### Minor Changes
+
+- Renamed alias for `--framework` argument from `--fw` to `-f` [#14](https://github.com/kubeshop/monokle-cli/issues/14).
+
+### Patch Changes
+
+- Updated dependencies.
+
 ## 0.3.0
 
 ### Minor Changes